PR review reconsiderations
